requirements.txt
django
channels[daphne]
django-ratelimit
python-dotenv
mysqlclient
django-jazzmin
google-generativeai
django-ratelimit